feat: create class pastTime for calculate pastTIme
- 과거 시간을 표시하기 위한 pastTime 생성
