+++ iOS/IssueTracker/IssueTracker/Issue/IssueList/View/IssueListViewCell.swift
@@ -26,41 +26,14 @@ class IssueListViewCell: UICollectionViewCell {
             UIColor(red: 0.8, green: 0.8, blue: 0.8, alpha: 0.4).cgColor
         containerView.layer.cornerRadius = 10
         title.text = issue.title
-        desc.text = "#\(issue.id) \(agoTime(from: issue.createdAt)) opened by \(issue.user.username)"
+        desc.text = "#\(issue.id) opened \(PastTime().agoTime(from: issue.createdAt)) by \(issue.user.username)"
         labelListView.labels = issue.labels
         if issue.state == 0 {
             infoImage.image = UIImage(systemName: "checkmark.circle")
             infoImage.tintColor = .systemRed
         }
     }
-    
-    func agoTime(from time: String) -> String {
-        let formatter = DateFormatter()
-        let format = "yyyy-MM-dd'T'HH:mm:ss.SSSZ"
-        formatter.dateFormat = format
-        guard let date = formatter.date(from: time)?.toGlobalTime() else { return ""}
-        let current = Date().toGlobalTime()
-        let diffTime = timeDiffMeasure(current: current, from: date)
-        return diffTime
-    }
-    
-    func timeDiffMeasure(current: Date, from privious: Date) -> String {
-        let s: Double = 60, m = s * s, h = m * 24, d = h * 365
-        let diff = current.timeIntervalSince1970 - privious.timeIntervalSince1970
-        switch diff {
-        case let second where second < s:
-            return "\(Int(second)) second ago"
-        case let minute where minute < m:
-            return "\(Int(minute / s)) minute ago"
-        case let hour where hour < h:
-            return "\(Int(hour / m)) hour ago"
-        case let day where day < d:
-            return "\(Int(day / h)) day ago"
-        default:
-            return ""
-        }
-    }
-
+ 
     override func prepareForReuse() {
         title.text = ""
         desc.text = ""
